Fix travel line detection
This is a bit annoying as the previous code _should_ have been correct if you match it against cura code. However, we encountered a very small line segment (1um) that was printed with 0 flow. As this was a lower flow target that _had_ to be reached the paths before/after had to decrease/increase in order to match this very low target speed.

CURA-10446