use target flow after undefined nozzle flow
At the start of a layer the flow in the nozzle is undefined,
we should keep this behaviour the same as we do now.
Just start from target flow.
If this gives problems now because we increase the speeds/flows
for these highspeed profiles, I propose that we add a specific
start-up flow for each feature.

Contributes to CURA-11044