"Add null check to prevent crash in GradualFlowPlugin"
A null check condition has been added to the 'usePlugin' method in GradualFlowPlugin.py to ensure 'activeMachine' is not None before accessing its 'extruderList'. Previously, if 'activeMachine' was null, it would lead to a crash while trying to access its properties. Therefore, this change improves the robustness of the code.

Fixes CURA-11144